refactor: extract manifest config modules
