+++ src/core/extension/manifest_action_config.rs
@@ -1,0 +1,124 @@
+use serde::{Deserialize, Serialize};
+use std::collections::HashMap;
+
+use super::manifest::{ActionType, BuiltinAction, HttpMethod};
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct RuntimeConfig {
+    /// Desktop app runtime type (python/shell/cli). CLI ignores this field.
+    #[serde(rename = "type", skip_serializing_if = "Option::is_none")]
+    pub runtime_type: Option<String>,
+
+    /// Shell command to execute when running the extension.
+    /// Template variables: {{entrypoint}}, {{args}}, {{extensionPath}}, plus project context vars.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub run_command: Option<String>,
+
+    /// Shell command to set up the extension (e.g., create venv, install deps).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub setup_command: Option<String>,
+
+    /// Shell command to check if extension is ready. Exit 0 = ready.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub ready_check: Option<String>,
+
+    /// Environment variables to set when running the extension.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub env: Option<HashMap<String, String>>,
+
+    /// Entry point file (used in template substitution).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub entrypoint: Option<String>,
+
+    /// Default args template (used in template substitution).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub args: Option<String>,
+
+    /// Default site for this extension (used by some CLI extensions).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub default_site: Option<String>,
+
+    /// Desktop app: Python dependencies to install.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub dependencies: Option<Vec<String>>,
+
+    /// Desktop app: Playwright browsers to install.
+    #[serde(rename = "playwrightBrowsers", skip_serializing_if = "Option::is_none")]
+    pub playwright_browsers: Option<Vec<String>>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct InputConfig {
+    pub id: String,
+    #[serde(rename = "type")]
+    pub input_type: String,
+    pub label: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub placeholder: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub default: Option<serde_json::Value>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub min: Option<i32>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub max: Option<i32>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub options: Option<Vec<SelectOption>>,
+    pub arg: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct SelectOption {
+    pub value: String,
+    pub label: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct OutputConfig {
+    pub schema: OutputSchema,
+    pub display: String,
+    pub selectable: bool,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct OutputSchema {
+    #[serde(rename = "type")]
+    pub schema_type: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub items: Option<HashMap<String, String>>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct ActionConfig {
+    pub id: String,
+    pub label: String,
+    #[serde(rename = "type")]
+    pub action_type: ActionType,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub endpoint: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub method: Option<HttpMethod>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub requires_auth: Option<bool>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub payload: Option<HashMap<String, serde_json::Value>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub command: Option<String>,
+    /// Builtin action type (Desktop app only). CLI parses but does not execute.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub builtin: Option<BuiltinAction>,
+    /// Column identifier for copy-column builtin action.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub column: Option<String>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct SettingConfig {
+    pub id: String,
+    #[serde(rename = "type")]
+    pub setting_type: String,
+    pub label: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub placeholder: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub default: Option<serde_json::Value>,
+}
